Добавил отрисовку коментов
